requirements.txt
Flask
opencv-python
numpy
Werkzeug
Pillow
torch
torchvision
transparent-background